feat: add Places component
